requirements.txt
antlr4-python3-runtime==4.7.2
llvmlite==0.43.0
pillow==11.2.1
pysimplegui==5.0.8.3
